+++ src/main/server/player/weapon.ts
@@ -8,17 +8,16 @@ export function useWeapon(player: alt.Player) {
     /**
      * Give and appy weapons to a player
      *
-     * @param {alt.IWeapon[]} weapons
+     * @param {(alt.IWeapon & { ammo: number })[])} weapons
      * @param {{ [hash: string]: number }} ammoData
      */
-    function apply(weapons: alt.IWeapon[], ammoData: { [hash: string]: number }) {
+    function apply(weapons: (alt.IWeapon & { ammo: number })[]) {
         const lastWeapon = player.currentWeapon;
 
         player.removeAllWeapons();
 
         for (let weapon of weapons) {
-            const ammo = ammoData[weapon.hash] ?? 0;
-            player.giveWeapon(weapon.hash, ammo, lastWeapon === weapon.hash);
+            player.giveWeapon(weapon.hash, weapon.ammo, lastWeapon === weapon.hash);
             player.setWeaponTintIndex(weapon.hash, weapon.tintIndex);
             for (let component of weapon.components) {
                 player.addWeaponComponent(weapon.hash, component);
@@ -62,23 +61,28 @@ export function useWeapon(player: alt.Player) {
         }
 
         const weapons = document.getField('weapons') ?? [];
-        const ammo = document.getField('ammo') ?? {};
-
         const index = weapons.findIndex((x) => x.hash === weaponHash);
         if (index >= 0) {
             weapons.splice(index, 1);
         }
 
-        if (ammo[weaponHash]) {
-            delete ammo[weaponHash];
+        await document.set('weapons', weapons);
+        sync();
+    }
+
+    /**
+     * Get all weapons that the character has currently
+     *
+     * @return
+     */
+    function getWeapons() {
+        const document = Rebar.document.character.useCharacter(player);
+        if (!document.get()) {
+            return [];
         }
 
-        await document.setBulk({
-            weapons,
-            ammo,
-        });
-
-        sync();
+        const weapons = document.getField('weapons') ?? [];
+        return weapons;
     }
 
     /**
@@ -96,16 +100,10 @@ export function useWeapon(player: alt.Player) {
         }
 
         const weapons = document.getField('weapons') ?? [];
-        const ammo = document.getField('ammo') ?? {};
 
-        weapons.push({ components: [], hash: alt.hash(model), tintIndex: 0 });
-        ammo[alt.hash(model)] = ammoCount;
+        weapons.push({ components: [], hash: alt.hash(model), tintIndex: 0, ammo: ammoCount });
 
-        await document.setBulk({
-            weapons,
-            ammo,
-        });
-
+        await document.set('weapons', weapons);
         sync();
     }
 
@@ -119,18 +117,21 @@ export function useWeapon(player: alt.Player) {
     async function addAmmo(model: string, ammoCount: number) {
         const document = Rebar.document.character.useCharacter(player);
         if (!document.get()) {
-            return;
+            return false;
         }
 
-        const ammo = document.getField('ammo') ?? {};
-        if (ammo[alt.hash(model)]) {
-            ammo[alt.hash(model)] += ammoCount;
-        } else {
-            ammo[alt.hash(model)] = ammoCount;
+        const modelHash = alt.hash(model);
+        const weapons = document.getField('weapons') ?? [];
+
+        const index = weapons.findIndex((x) => x.hash === modelHash);
+        if (index <= -1) {
+            return false;
         }
 
-        player.setWeaponAmmo(model, ammo[alt.hash(model)]);
-        document.set('ammo', ammo);
+        weapons[index].ammo += ammoCount;
+        player.setWeaponAmmo(model, weapons[index].ammo);
+        document.set('weapons', weapons);
+        return true;
     }
 
     /**
@@ -142,31 +143,38 @@ export function useWeapon(player: alt.Player) {
             return;
         }
 
-        const ammo: { [key: string]: number } = {};
-        for (let weapon of player.weapons) {
-            ammo[weapon.hash] = player.getAmmo(weapon.hash);
-        }
-
-        document.setBulk({
-            weapons: player.weapons,
-            ammo: ammo,
-        });
-    }
-
-    function saveAmmo() {
-        const document = Rebar.document.character.useCharacter(player);
-        if (!document.get()) {
+        const weapons = document.getField('weapons');
+        if (!weapons || weapons.length <= 0) {
             return;
         }
 
-        const ammo: { [key: string]: number } = {};
         for (let weapon of player.weapons) {
-            ammo[weapon.hash] = player.getAmmo(weapon.hash);
+            const index = weapons.findIndex((x) => x.hash === weapon.hash);
+            if (index <= -1) {
+                continue;
+            }
+
+            const { ammoTypeHash } = alt.getWeaponModelInfoByHash(weapon.hash);
+            weapons[index].ammo = player.getAmmo(ammoTypeHash);
         }
 
-        document.set('ammo', ammo);
+        document.set('weapons', weapons);
     }
 
+    /**
+     * Calls the `save` function
+     *
+     * @deprecated
+     */
+    function saveAmmo() {
+        save();
+    }
+
+    /**
+     * Synchronize the weapons the player currently has
+     *
+     * @return
+     */
     function sync() {
         const document = useCharacter(player);
         const data = document.get();
@@ -174,7 +182,7 @@ export function useWeapon(player: alt.Player) {
             return;
         }
 
-        apply(data.weapons, data.ammo);
+        apply(data.weapons);
     }
 
     return {
@@ -183,6 +191,7 @@ export function useWeapon(player: alt.Player) {
         apply,
         clear,
         clearWeapon,
+        getWeapons,
         save,
         saveAmmo,
         sync,